% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjca.R
\name{dict_slice}
\alias{dict_slice}
\title{Extract counts from a hierarchical dictionary}
\usage{
dict_slice(df, level = 1, subtree = TRUE)
}
\arguments{
\item{df}{the dictionary output}

\item{level}{which level of the dictionary to get counts for}

\item{subtree}{whether to show a full path from the dictionary root (default: FALSE)}
}
\value{
dictionary results without super or sub category counts
}
\description{
Takes a complete dictionary report and constructs a
valid contingency table by slicing the dictionary at `level`.
Uses \code{_} as a level marker in the column names.
}

