% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjca.R
\name{get_encoding}
\alias{get_encoding}
\title{Determine default encoding}
\usage{
get_encoding()
}
\value{
string representation of a file encoding
}
\description{
Determines default file encoding by parsing the LC_COLLATE locale value
}

