% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjca.R
\name{get_locale}
\alias{get_locale}
\title{Determine default locale}
\usage{
get_locale()
}
\value{
ISO string representation of a locale
}
\description{
Determines default locale by parsing the LC_COLLATE locale value
}

