% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjca.R
\name{insert_defaults}
\alias{insert_defaults}
\title{Fill in default values for a list}
\usage{
insert_defaults(lst, defs)
}
\arguments{
\item{lst}{List to be augmented}

\item{defs}{Default entry values for \code{lst} if it doesn't already have them}
}
\value{
an augmented list
}
\description{
Adds entries to a list where necessary.
}

