% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjca.R
\name{jca_conc}
\alias{jca_conc}
\title{Drive the concordancer}
\usage{
jca_conc(files, ...)
}
\arguments{
\item{files}{a list of file and folder names}

\item{...}{extra arguments to control the process. See Details.}
}
\value{
a data.frame containing
}
\description{
Drive the concordancer and return the concordance as a data.frame.
See details for possible extra arguments.
}
\details{
Possible extra arguments are
\itemize{
 \item{\code{locale}: }{locale in ISO format, e.g. "en_US".
   Defaults to whatever \code{get_locale} returns}
 \item{\code{encoding}: }{the file encoding of the \code{files}. Defaults to the
  whatever \code{get_encoding} returns}
 \item{\code{output}: }{the folder that the results will land in before being served
  back as a data.frame.  Defaults to a temporary folder}
 \item{\code{progress}: }{whether to show the progress of the java code as it goes
 through the \code{files}}
 \item{\code{dictionary}: }{a content analysis dictionary in ykd or vbpro format}
 \item{\code{category}: }{optional name of a category within \code{dictionary}}
 \item{\code{pattern}: }{a word or phrase, possible containing wildcard character '*'}
 \item{\code{window}: }{how many words either side to show}
}

Either \code{dictionary} or \code{pattern} must be provided.  If \code{dictionary} is
provided then \code{category} may also be provided.  If it is not then the concordance
is of all the patterns anywhere in the dictionary.

This function also dumps the location of the temporary folder where the
results landed to standard error in case you want the original html data file
from which the data.frame was extracted.

Note: Concordances can get big.  This function does not try to be efficient.
}

