% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjca.R
\name{jca_desc}
\alias{jca_desc}
\title{Drive the document describer}
\usage{
jca_desc(files, ...)
}
\arguments{
\item{files}{a list of file and folder names}

\item{...}{extra arguments to control the process. See Details.}
}
\value{
a data.frame containing brief document descriptions
}
\description{
Drive the document describer.  See details for possible extra arguments.
}
\details{
Optional arguments are
\itemize{
 \item{\code{locale}: }{Locale in ISO format, e.g. "en_US".
   Defaults to whatever \code{get_locale} returns}
 \item{\code{encoding}: }{the file encoding of the \code{files}. Defaults to the
  whatever \code{get_encoding} returns}
 \item{\code{output}: }{the file that the results will land in before being served
  back as a data.frame.  Defaults to a temporary file}
 \item{\code{progress}: }{Whether to show the progress of the java code as it goes
 through the \code{files}}
}
}

