% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjca.R
\name{jca_line}
\alias{jca_line}
\title{Drive the single liner}
\usage{
jca_line(files, filename, ...)
}
\arguments{
\item{files}{a list of file and folder names}

\item{filename}{the single file you want the contents of \code{files} in}

\item{...}{extra arguments to control the process. See Details.}
}
\description{
Puts a set of documents into a single file in one of the ways Mallet likes.
This function creates lines of the form
}
\details{
\code{folder-name | NoLabel} tab \code{filename} tab \code{file contents...}

This is the format called 'one instance per file' on the Mallet pages
(http://mallet.cs.umass.edu/import.php).  If you provide folder names to
the \code{files} argument the first line element will be the folder name
and the second element will be the names of file within it. If you provide filenames
individually then the first line element will be 'NoLabel' and the second element
will be the filename.  You can mix these.

Note that you cannot provide folders of folders of files.  The function looks for
files and folders in \code{files}.  For each of the folders it looks for files and adds
them.  Then it stops looking.  Also, files or folders with names beginning
with '.' will be ignored.

Optional arguments are
\itemize{
 \item{\code{locale}: }{Locale in ISO format, e.g. "en_US".
   Defaults to whatever \code{get_locale} returns}
 \item{\code{encoding}: }{the file encoding of the \code{files}. Defaults to the
 result of \code{get_encoding}}
}
}

