% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjca.R
\name{read_mtx}
\alias{read_mtx}
\title{Read the a word frequency matrix in Matrix Market format}
\usage{
read_mtx(folder)
}
\arguments{
\item{folder}{folder containing the data}
}
\description{
Returns an sparse Matrix object from a folder generated by \code{jca_word}.
The folder is expected to contain 'data.mtx', a word count matrix
with documents as rows in Matrix Market format, a file of row names
'documents.csv', and column names 'words.csv'.  Uses the Matrix packages \code{readMM}
underneath.
}

