% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjca.R
\docType{package}
\name{rjca}
\alias{rjca}
\alias{rjca-package}
\title{rjca: A package for driving jca java code.}
\description{
The rjca package provides functions to count words, count categories
using a content analysis dictionary, and generate concordances for words
or categories.
}

