% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjca.R
\name{tidy_conc}
\alias{tidy_conc}
\title{Turn jca_conc output into something that looks like a concordance}
\usage{
tidy_conc(df)
}
\arguments{
\item{df}{A three column data.frame with incomplete document names}
}
\value{
a pretty data.frame that looks kind of like a concordance
}
\description{
Constructs a data frame with document names in the first column and
string-aligned concordance lines in the second column.  Blank document
names are filled in as necessary.  Used to format
output in \code{jca_conc}.
}

