% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjca.R
\name{jca_line}
\alias{jca_line}
\title{Drive the single liner}
\usage{
jca_line(files, filename, ...)
}
\arguments{
\item{files}{a list of file and folder names}

\item{filename}{the single file you want the contents of \code{files} in}

\item{...}{extra arguments to control the process. See Details.}
}
\description{
Puts a set of documents into a single file, in the way Mallet likes.
}
\details{
Optional arguments are
\itemize{
 \item{\code{locale}: }{Locale in ISO format, e.g. "en_US".
   Defaults to whatever \code{get_locale} returns}
 \item{\code{encoding}: }{the file encoding of the \code{files}. Defaults to the
 result of \code{get_encoding}}
}
}

