% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjca.R
\name{jca_word}
\alias{jca_word}
\title{Drive the word counter}
\usage{
jca_word(files, ...)
}
\arguments{
\item{files}{a list of file and folder names}

\item{...}{extra arguments to control the process. See Details.}
}
\value{
a sparse Matrix containing the word counts
}
\description{
Drive the word counter and return a sparse matrix containing word counts.
Optional arguments are
\itemize{
 \item{\code{locale}: }{Locale in ISO format, e.g. "en_US".
   Defaults to whatever \code{get_locale} returns}
 \item{\code{encoding}: }{the file encoding of the \code{files}. Defaults to the
  whatever \code{get_encoding} returns}
 \item{\code{output}: }{the folder that the results will land in before being served
  back as a data.frame.  Defaults to a temporary folder}
 \item{\code{progress}: }{Whether to show the progress of the java code as it goes
 through the \code{files}}
}
This function also dumps the location of the temporary folder where the
results landed to
standard error in case you want the original csv data file and yoshikoder dictionary
translation.
}

